+++ apibuilder4elastic/custom_flow_nodes/api-builder-plugin-axway-api-management/test/test-getCustomProperties.js
@@ -1,0 +1,39 @@
+const { expect } = require('chai');
+const { MockRuntime } = require('@axway/api-builder-test-utils');
+const getPlugin = require('../src');
+const path = require('path');
+const fs = require('fs');
+const nock = require('nock');
+const envLoader = require('dotenv');
+const decache = require('decache');
+
+describe('Test Get custom properties from API-Manager', () => {
+	let plugin;
+	let flowNode;
+
+	// Loads environment variables from .env if the file exists
+	const envFilePath = path.join(__dirname, '.env');
+	if (fs.existsSync(envFilePath)) {
+		delete process.env.API_MANAGER; // Otherwise it is not overwritten
+		envLoader.config({ path: envFilePath });
+	}
+	// Delete the cached module 
+	decache('../config/axway-api-utils.default.js');
+	decache('../../../conf/elasticsearch.default.js');
+	var pluginConfig = require('../config/axway-api-utils.default.js').pluginConfig['api-builder-plugin-axway-api-management'];
+
+	beforeEach(async () => {
+		plugin = await MockRuntime.loadPlugin(getPlugin,pluginConfig);
+		plugin.setOptions({ validateOutputs: true });
+		flowNode = plugin.getFlowNode('axway-api-management');
+	});
+
+	describe('#getCustomProperties', () => {
+		it('should not return custom properties if feature is disabled', async () => {
+			const { value, output } = await flowNode.getCustomPropertiesConfig({ disableCustomProperties: true });
+
+			expect(value).to.deep.equal({});
+			expect(output).to.equal('next');
+		});
+	});
+});